Create Listing4.10-Writing-Two-Versions-of-a-Function
The first call includes two strings. The compiler knows to use the first version.
The second call includes two integers. The compiler knows to use the second version.
This process of writing two versions of the same function is called overloading the function. 
When overloading a function, the parameters of a function differ.